pyproject.toml
[project]
name = "python-claude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.3",
    "langchain-mongodb>=0.11.0",
    "langchain-voyageai>=0.3.2",
    "langgraph>=1.0.9",
    "motor>=3.7.1",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
    "voyageai>=0.3.7",
]
